Move part of MO_Reader to an abstract class
